Re-scaffold database

+++ WebApps/ShoppingList/ShoppingList/Database/Entities/ItemOnShoppingList.cs
@@ -8,7 +8,7 @@ namespace ShoppingList.Database.Entities
         public uint ItemOnShoppingListId { get; set; }
         public uint ShoppingListId { get; set; }
         public uint ItemId { get; set; }
-        public decimal? Quantity { get; set; }
+        public uint? Quantity { get; set; }
 
         public virtual Item Item { get; set; }
         public virtual ShoppingList ShoppingList { get; set; }
